Gp user part added completly
