--- src/groups/groupUser/gpUser.dto.ts
@@ -1,20 +1,0 @@
-import { IsUUID, IsObject, ValidateNested, IsString, IsArray, isArray } from 'class-validator';
-import { Type } from 'class-transformer';
-
-
-class UserIdDto {
-  @IsString()
-  userId: string;
-}
-class GPDTO{
-  groupId : string
-  @IsArray()
-  @IsString()
-  userIds : UserIdDto[]
-}
-export class addUserDTO{
-  @IsArray()
-  @Type(()=>GPDTO)
-  members : GPDTO[]
-}
-
