+++ src/groups/groups.module.ts
@@ -5,12 +5,10 @@ import { GroupsController } from './groups.controller';
 import { GroupsService } from './groups.service';
 import { createGroupDTO } from './groups.dto';
 import { GroupUserService } from './groupUser/gpUser.service';
-import { GroupUserController } from './groupUser/gpUser.controller';
-import { addUserDTO, deleteMemberDTO } from './groupUser/gpUser.dto';
 
 @Module({
-  controllers: [GroupUserController,GroupsController],
-  providers: [GroupUserService, GroupsService,PrismaService,createGroupDTO, addUserDTO, deleteMemberDTO],
-  exports : [createGroupDTO, addUserDTO, deleteMemberDTO]
+  controllers: [GroupsController],
+  providers: [GroupUserService, GroupsService, PrismaService, createGroupDTO],
+  exports: [createGroupDTO]
 })
-export class GroupModule {}
+export class GroupModule { }
